added functionality for exporting of alignments on upload
